Revert "Add cgroup-bin dependency to our Ubuntu package"
This reverts commit efc4b0f7035497371182761fb56f8467bfbb5772.

After re-reading the documentation: "The Recommends field should list packages that would be found together with this one in all but unusual installations."

Thus, "Recommends" is an acceptable place for this dep, and anyone disabling that gets to keep the pieces.

The main crux of why this needs to be reverted is because it breaks Debian completely because "lxc" and "cgroup-bin" can't be installed concurrently.