Makefile: don't allocate pseudo-tty if not interactive
In cases where we are running non-interactively (e.g. on Jenkins), we
shouldn't try to allocate a TTY, which would fail. This patch makes the
flag `-t` be dependent on whether the shell session is interactive or
not.

When it is interactive, it's important to have `-t` so that the user can
pass signals through to the process.

We also remove the DOCKER_RUN_DOCS and the GITCOMMIT variables, which
are no longer in use in this Makefile (seems like the related targets
that used it were migrated to docs/Makefile).

Signed-off-by: Jonathan Lebon <jlebon@redhat.com>